refactor(Scores): keys noramlization now in StatsTable component
