+++ src/Ml.Cli.WebApp/ClientApp/src/Comparison/StatsTable.js
@@ -1,5 +1,5 @@
 ﻿import React, {useState} from "react";
-import {totalCompletenessByKey, totalScores} from "./score";
+import {normalizeKeys, totalScores} from "./score";
 import {filterItems} from './TableResult';
 import ExcelExport from "./ExcelExport";
 import {CheckboxInput, CheckboxModes} from "@axa-fr/react-toolkit-form-input-checkbox";
@@ -53,46 +53,6 @@ export const calcTotalCompleteness = (levenshteinResults, numberItems) => {
     return {left: totalLeftCompleteness, right: totalRightCompleteness};
 };
 
-const updateKeysDict = (keysDict, newKey, value) => {
-    let objectValues = keysDict[newKey];
-    objectValues.score += value.score;
-    objectValues.score_ok += value.score_ok;
-    objectValues.score_ko += value.score_ko;
-    objectValues.total += value.total;
-    objectValues.completeness_left.ok += value.completeness_left.ok;
-    objectValues.completeness_left.ko += value.completeness_left.ko;
-    objectValues.completeness_right.ok += value.completeness_right.ok;
-    objectValues.completeness_right.ko += value.completeness_right.ko;
-    keysDict[newKey] = objectValues;
-    
-    const percentagesOK = ((objectValues.score_ok + objectValues.score_ko) === 0 ? "-" : `${Math.round((objectValues.score_ok / (objectValues.score_ok + objectValues.score_ko))*100*100)/100} %`);
-    const tempLeftSide = {completeness_ok: objectValues.completeness_left.ok, completeness_ko: objectValues.completeness_left.ko};
-    const tempRightSide = {completeness_ok: objectValues.completeness_right.ok, completeness_ko: objectValues.completeness_right.ko};
-    const completenessOKLeft = totalCompletenessByKey(tempLeftSide);
-    const completenessOKRight = totalCompletenessByKey(tempRightSide);
-    
-    objectValues.percentages.ok = percentagesOK;
-    objectValues.percentages.completeness_ok_left = `${completenessOKLeft} %`;
-    objectValues.percentages.completeness_ok_right = `${completenessOKRight} %`;
-    
-    return keysDict;
-};
-
-const normalizeKeys = (levenshteinResult, keys) => {
-    let newKeysDict = {};
-    const regex = /[0-9]+_/ig;
-    keys.forEach(key => {
-        const newKey = key.replace(regex, "");
-        if(newKeysDict.hasOwnProperty((newKey))){
-            newKeysDict = updateKeysDict(newKeysDict, newKey, levenshteinResult[key]);
-        }
-        else{
-            newKeysDict[newKey] = levenshteinResult[key];
-        }
-    });
-    return newKeysDict;
-};
-
 const StatusCode = ({statusCodes}) => {
     const statuses =  Object.keys(statusCodes).map(function(key) {
         const value = statusCodes[key];
@@ -108,12 +68,8 @@ const StatusCode = ({statusCodes}) => {
     return <>{statuses}</>
 }
 
-const Scores = ({levenshteinResults, isMerging}) => {
+const Scores = ({levenshteinResults}) => {
     let keys = Object.keys(levenshteinResults);
-    if(isMerging){
-        levenshteinResults = normalizeKeys(levenshteinResults, keys);
-        keys = Object.keys(levenshteinResults);
-    }
     const ScoreItems =  keys.map(function(key) {
         const keyValue = levenshteinResults[key];
         return <div className="stats__results" key={key}>
@@ -174,7 +130,10 @@ const StatsTable = ({state, setState, items}) => {
     const filteredItems = statsState.areErrorsRemoved ?
         items.filter(item => item.left.StatusCode === 200 && item.right.StatusCode === 200) :
         items;
-    const levenshteinResults = totalScores(filteredItems, ["document_id", "document_type"]);
+    let levenshteinResults = totalScores(filteredItems, ["document_id", "document_type"]);
+    if(statsState.isMerging){
+        levenshteinResults = normalizeKeys(levenshteinResults, Object.keys(levenshteinResults));
+    }
     const totalTimeMs = computeTotalTimeMs(filteredItems);
     const statusCodes = orderByStatusCode(filteredItems);
     const totalCompleteness = calcTotalCompleteness(levenshteinResults, filteredItems.length);
@@ -258,7 +217,6 @@ const StatsTable = ({state, setState, items}) => {
                         <StatusCode statusCodes={statusCodes}  />
                         <Scores
                             levenshteinResults={levenshteinResults}
-                            isMerging={statsState.isMerging}
                         />
                     </>
                 )}
